// application/code/book-my-show/src/application/models/aux-data/Country.Interface.ts
/**
 * Author: Brian Brown
 * Date: June 22nd, 2020
 * Description: Used to hold countries for selection
 *
 * Change Log:
 *
 * Date         Developer           Description
 * ---------    -----------------   -------------------------------------------------
 *
 */
interface Country {
    countryCode: string;
    countryName: string;
}

export { Country };